Groundwork for arch & platform distinction
This is some preliminary work to make a distinction between an "arch" (i.e. typically a MCU family, such as the RP2040) and a "platform", which typically refers to boards and such hardware configurations depending on an arch.

The arch directory groups up common code that is entry points for the hardware.
Each arch, if it should support platforms, can expose their own API similar to the HAL.
Moreover, by convention, it may also defer some of the HAL implementation to the platform - e.g. button I/O may involve common platform code; but its use is most certainly board-specific.